feat: add reset states if player is hit
If the player is hit by an enemy or an enemy bullet the game will now reset
the enemies back to the top of the screen and the player to his starting
position. If the hit causes the player to have 0 lives, the game will reset
to its starting state.