fix: fix source of major slowdown
After a certain point, the game would slow down to the point it pretty much
did not run. This was caused by the call to Rect(), used to draw a rectangle
to the canvas to fill in a black background, being in the loop. This meant
the rectangles would stack up without deleting. Now only the fill function
remains in the loop, as is necessary.